# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Compta EI API
  description: API REST pour la gestion comptable d'une Entreprise Individuelle - 45 endpoints testés et fonctionnels
  version: 2.0.0
  contact:
    name: Compta EI
    email: contact@monentreprise.fr
servers:
  - url: http://localhost:3000/api
    description: Serveur de développement
  - url: https://api.compta-ei.fr/api
    description: Serveur de production
...
paths:
...
  /tva/calcul/{debut}/{fin}:
    get:
      tags:
        - TVA
      summary: Calculate TVA for period
      description: Calcule la TVA pour une période donnée
      parameters:
        - name: debut
          in: path
          required: true
          schema:
            type: string
            format: date
          example: '2025-01-01'
        - name: fin
          in: path
          required: true
          schema:
            type: string
            format: date
          example: '2025-12-13'
      responses:
        '200':
          description: Calcul de la TVA
          content:
            application/json:
              schema:
                type: object
                properties:
                  tvaCollectee:
                    type: number
                    format: float
                  tvaDeductible:
                    type: number
                    format: float
                  tvaADecaisser:
                    type: number
                    format: float
                  periodeDebut:
                    type: string
                    format: date
                  periodeFin:
                    type: string
                    format: date
...
  /tva/detail-taux/{debut}/{fin}:
    get:
      tags:
        - TVA
      summary: Get TVA detail by rate
      description: Détail de la TVA par taux pour une période
      parameters:
        - name: debut
          in: path
          required: true
          schema:
            type: string
            format: date
          example: '2025-01-01'
        - name: fin
          in: path
          required: true
          schema:
            type: string
            format: date
          example: '2025-12-31'
      responses:
        '200':
          description: Détail TVA par taux
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    taux:
                      type: string
                    tva_collectee:
                      type: string
                    tva_deductible:
                      type: string
...
  /immobilisations:
    get:
      tags:
        - Immobilisations
      summary: Get all fixed assets
      description: Récupère toutes les immobilisations
      responses:
        '200':
          description: Liste des immobilisations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Immobilisation'
    post:
      tags:
        - Immobilisations
      summary: Create fixed asset
      description: Crée une nouvelle immobilisation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImmobilisationInput'
      responses:
        '201':
          description: Immobilisation créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Immobilisation'
  /immobilisations/{id}:
    put:
      tags:
        - Immobilisations
      summary: Update fixed asset
      description: Met à jour une immobilisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImmobilisationInput'
      responses:
        '200':
          description: Immobilisation mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Immobilisation'
    delete:
      tags:
        - Immobilisations
      summary: Delete fixed asset
      description: Supprime une immobilisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Immobilisation supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Immobilisation supprimée avec succès"
  /immobilisations/{id}/amortissements:
    get:
      tags:
        - Immobilisations
      summary: Get asset depreciations
      description: Récupère les amortissements d'une immobilisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste des amortissements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Amortissement'
  /immobilisations/{id}/amortissement/{annee}:
    get:
      tags:
        - Immobilisations
      summary: Calculate depreciation for year
      description: Calcule l'amortissement pour une année donnée
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: annee
          in: path
          required: true
          schema:
            type: integer
          example: 2025
      responses:
        '200':
          description: Calcul de l'amortissement
          content:
            application/json:
              schema:
                type: object
                properties:
                  immobilisation_id:
                    type: integer
                  annee:
                    type: integer
                  montant:
                    type: number
                  valeur_nette_comptable:
                    type: number
...
components:
  schemas:
...
    Immobilisation:
      type: object
      properties:
        id:
          type: integer
        libelle:
          type: string
          example: "Ordinateur portable"
        type:
          type: string
          example: "Materiel informatique"
        date_acquisition:
          type: string
          format: date-time
        valeur_acquisition:
          type: string
        duree_amortissement:
          type: integer
        methode_amortissement:
          type: string
          enum: [lineaire, degressif]
        taux_amortissement:
          type: string
        valeur_residuelle:
          type: string
        compte_immobilisation:
          type: string
        compte_amortissement:
          type: string
        en_service:
          type: boolean
        date_cession:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ImmobilisationInput:
      type: object
      required:
        - libelle
        - type
        - date_acquisition
        - valeur_acquisition
        - duree_amortissement
      properties:
        libelle:
          type: string
        type:
          type: string
        date_acquisition:
          type: string
          format: date
        valeur_acquisition:
          type: number
        duree_amortissement:
          type: integer
        methode_amortissement:
          type: string
          enum: [lineaire, degressif]
          default: lineaire
        taux_amortissement:
          type: number
        valeur_residuelle:
          type: number
          default: 0
        compte_immobilisation:
          type: string
        compte_amortissement:
          type: string
        en_service:
          type: boolean
          default: true
        notes:
          type: string
    Amortissement:
      type: object
      properties:
        id:
          type: integer
        immobilisation_id:
          type: integer
        annee:
          type: integer
        montant_amortissement:
          type: string
        amortissement_cumule:
          type: string
        valeur_nette_comptable:
          type: string
        ecriture_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
...